Multiple phenology changes
(issue #165) Added a rescale_phenology flag. When TRUE sxwphen.in and sxwprod_v2.in files will be generated for each climate scenario.

(issue #159) I also ensured the phenology values sum to 1 across rows.

I also rounded all of the values we write to 9 decimal places, to make sure we aren't writing scientific notation to the files (which c might or might not be able to read).

+++ R_program/CallSTEPWAT2.R
@@ -45,24 +45,26 @@ foreach (g = 1:length(GCM)) %dopar% { # loop through all the GCMs
         weather.dir<-paste(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw/Input/",sep="")
         weather.dir2<-paste(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw/Input/randomdata/",sep="")
         
-        ################ Copy phen and prod files generated for current conditions #################
-        setwd(dist.directory)
-        # Copy the phenology file generated for this specific GCM, RCP and YEAR.
-        system(paste0("cp ", "sxwphen.", GCM[g], ".in", " ", directory,
-                      "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-        # Copy the prod file generated for this specific GCM, RCP and YEAR.
-        system(paste0("cp ", "sxwprod_v2.", GCM[g], ".in", " ", directory,
-                      "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-        # Move to the SXW inputs for STEPWAT2
-        setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-        # Remove the old phenology file
-        system("rm sxwphen.in")
-        # Remove the old prod file
-        system("rm sxwprod_v2.in")
-        # Rename the phenology file to the name recognized by STEPWAT2
-        system(paste0("mv ", "sxwphen.", GCM[g], ".in", " sxwphen.in"))
-        # Rename the prod file to the name recognized by STEPWAT2
-        system(paste0("mv ", "sxwprod_v2.", GCM[g], ".in", " sxwprod_v2.in"))
+        if(rescale_phenology){
+          ################ Copy phen and prod files generated for current conditions #################
+          setwd(dist.directory)
+          # Copy the phenology file generated for this specific GCM, RCP and YEAR.
+          system(paste0("cp ", "sxwphen.", GCM[g], ".in", " ", directory,
+                        "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+          # Copy the prod file generated for this specific GCM, RCP and YEAR.
+          system(paste0("cp ", "sxwprod_v2.", GCM[g], ".in", " ", directory,
+                        "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+          # Move to the SXW inputs for STEPWAT2
+          setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+          # Remove the old phenology file
+          system("rm sxwphen.in")
+          # Remove the old prod file
+          system("rm sxwprod_v2.in")
+          # Rename the phenology file to the name recognized by STEPWAT2
+          system(paste0("mv ", "sxwphen.", GCM[g], ".in", " sxwphen.in"))
+          # Rename the prod file to the name recognized by STEPWAT2
+          system(paste0("mv ", "sxwprod_v2.", GCM[g], ".in", " sxwprod_v2.in"))
+        }
         
         #Copy the weather data into the randomdata folder, commenting out creation of weather.in files as default so rSFSTEP2
         #uses only weather data generated from the markov weather generator but retain this code if one wants to create and copy 30 years of 
@@ -176,24 +178,26 @@ foreach (g = 1:length(GCM)) %dopar% { # loop through all the GCMs
               weath.read<-paste(assembly_output,"Site_",s,"_hybrid-delta.",y,".",r,".",GCM[g], sep="")
             }
             
-            ############## Move phen and prod files generated for this specific GCM x RCP x YEARS combination #############
-            setwd(dist.directory)
-            # Copy the phenology file generated for this climate into STEPWAT2
-            system(paste0("cp ", "sxwphen.", downscaling_method, ".", r, ".", GCM[g], ".in", " ", directory,
-                          "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-            # Copy the prod file generated for this climate into STEPWAT2
-            system(paste0("cp ", "sxwprod_v2.", downscaling_method, ".", r, ".", GCM[g], ".in", " ", directory,
-                          "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-            # Move into the sxw inputs folder for STEPWAT2
-            setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
-            # Remove the old phenology file
-            system("rm sxwphen.in")
-            # Remove the old prod file
-            system("rm sxwprod_v2.in")
-            # Rename the new phenology file to the name recognized by STEPWAT2
-            system(paste0("mv ", "sxwphen.", downscaling_method, ".", r, ".", GCM[g], ".in", " sxwphen.in"))
-            # Rename the new prod file to the name recognized by STEPWAT2
-            system(paste0("mv ", "sxwprod_v2.", downscaling_method, ".", r, ".", GCM[g], ".in", " sxwprod_v2.in"))
+            if(rescale_phenology){
+              ########## Move phen and prod files generated for this specific GCM x RCP x YEARS combination #########
+              setwd(dist.directory)
+              # Copy the phenology file generated for this climate into STEPWAT2
+              system(paste0("cp ", "sxwphen.", downscaling_method, ".", r, ".", GCM[g], ".in", " ", directory,
+                            "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+              # Copy the prod file generated for this climate into STEPWAT2
+              system(paste0("cp ", "sxwprod_v2.", downscaling_method, ".", r, ".", GCM[g], ".in", " ", directory,
+                            "Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+              # Move into the sxw inputs folder for STEPWAT2
+              setwd(paste0(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw"))
+              # Remove the old phenology file
+              system("rm sxwphen.in")
+              # Remove the old prod file
+              system("rm sxwprod_v2.in")
+              # Rename the new phenology file to the name recognized by STEPWAT2
+              system(paste0("mv ", "sxwphen.", downscaling_method, ".", r, ".", GCM[g], ".in", " sxwphen.in"))
+              # Rename the new prod file to the name recognized by STEPWAT2
+              system(paste0("mv ", "sxwprod_v2.", downscaling_method, ".", r, ".", GCM[g], ".in", " sxwprod_v2.in"))
+            }
             
             #Identify the directory the weather will be pasted into   
             weather.dir<-paste(directory,"Stepwat.Site.",s,".",g,"/testing.sagebrush.master/Stepwat_Inputs/Input/sxw/Input/",sep="")
